pyproject.toml
[tool.poetry]
name = "sales-report-bot-v2"
version = "0.1.0"
description = ""
authors = ["royce <royceinbox@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.6"
pytz = "^2023.3.post1"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pre-commit = "^3.4.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
